[front] refactor: keep sub-agent fetch in reader
