+++ front/lib/resources/agent_message_consumption_item_resource.ts
@@ -15,7 +15,6 @@ import type { ModelId } from "@app/types/shared/model_id";
 import type { Result } from "@app/types/shared/result";
 import { Err } from "@app/types/shared/result";
 import { assertNever } from "@app/types/shared/utils/assert_never";
-import { removeNulls } from "@app/types/shared/utils/general";
 import assert from "assert";
 import type { Attributes, CreationAttributes, Transaction } from "sequelize";
 import { Op, QueryTypes } from "sequelize";
@@ -816,9 +815,9 @@ export class AgentMessageConsumptionItemResource extends BaseResource<AgentMessa
   }
 
   /**
-   * Resolves one public message identity and its direct sub-agent subtree facts. Keeping that
-   * resolution here prevents persistence traversal and numeric IDs from leaking into the read
-   * module or route.
+   * Resolves one public message identity to the persisted facts needed by the consumption reader.
+   * Keeping that resolution here prevents Sequelize models and numeric IDs from leaking into the
+   * read module or route.
    */
   static async fetchMessageConsumptionFacts(
     auth: Authenticator,
@@ -836,11 +835,6 @@ export class AgentMessageConsumptionItemResource extends BaseResource<AgentMessa
     dustRunIds: string[];
     items: AgentMessageConsumptionItemResource[];
     actions: AgentMCPActionResource[];
-    directSubAgents: {
-      action: AgentMCPActionResource;
-      agentConfigurationId: string | null;
-      subtreeBilledCredits: number;
-    }[];
   } | null> {
     const messageRes = await conversation.getMessageById(auth, agentMessageId);
     if (messageRes.isErr() || !messageRes.value.agentMessage) {
@@ -856,62 +850,11 @@ export class AgentMessageConsumptionItemResource extends BaseResource<AgentMessa
       AgentMCPActionResource.listByAgentMessageIds(auth, [agentMessage.id]),
     ]);
 
-    const directSubAgentRoots = removeNulls(
-      actions.map((action) => {
-        const childConversationId = action.getRunAgentChildConversationId();
-        return childConversationId && childConversationId !== conversation.sId
-          ? { action, childConversationId }
-          : null;
-      })
-    );
-    const childConversations = await ConversationResource.fetchByIds(
-      auth,
-      [
-        ...new Set(
-          directSubAgentRoots.map(
-            ({ childConversationId }) => childConversationId
-          )
-        ),
-      ],
-      { includeDeleted: true }
-    );
-    const childFactsByConversationId = new Map<
-      string,
-      { agentConfigurationId: string | null; subtreeBilledCredits: number }
-    >();
-    for (const childConversation of childConversations) {
-      const { messages } = await this.fetchConversationConsumptionFacts(auth, {
-        conversation: childConversation,
-        maxAttributionVersion,
-      });
-      const rootMessage = messages.find(
-        (message) => message.conversationId === childConversation.sId
-      );
-      childFactsByConversationId.set(childConversation.sId, {
-        agentConfigurationId: rootMessage?.agentConfigurationId ?? null,
-        subtreeBilledCredits: messages.reduce(
-          (total, message) => total + (message.billedCredits ?? 0),
-          0
-        ),
-      });
-    }
-
     return {
       billedCredits: agentMessage.costCredits,
       dustRunIds: [...new Set(agentMessage.runIds ?? [])],
       items,
       actions,
-      directSubAgents: directSubAgentRoots.map(
-        ({ action, childConversationId }) => {
-          const childFacts =
-            childFactsByConversationId.get(childConversationId);
-          return {
-            action,
-            agentConfigurationId: childFacts?.agentConfigurationId ?? null,
-            subtreeBilledCredits: childFacts?.subtreeBilledCredits ?? 0,
-          };
-        }
-      ),
     };
   }
